% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_parameters.R
\name{formatParameters}
\alias{formatParameters}
\title{parameters_info is a list with elements beta, gamma, theta
each needs columns var, name, theta needs log}
\usage{
formatParameters(x, parameters_info)
}
\description{
parameters_info is a list with elements beta, gamma, theta
each needs columns var, name, theta needs log
}
