% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{getEffect}
\alias{getEffect}
\title{Helper to visualize results (of an \code{hlnm} fit) associated with a given variable.}
\usage{
getEffect(fit, exposure_var, group_var, group, values, ref_values)
}
\arguments{
\item{fit}{Output of an \code{hlnm} call.}

\item{exposure_var}{character indicating the variable of interest.}

\item{group_var}{character indicating the variable upon which to segment the results (should match the \code{group_var}
used in the hierarchical models of the \code{exposure_var}). Can be a vector, but this might lead to problems.}

\item{group}{character vector indicating specific group(s) (possible value(s) of the \code{group_var} variable) of interest.}

\item{values}{Numeric values at which to evaluate the \code{exposure_var} (set to the range spanned by the knots, if applicable).}

\item{ref_values}{Named list indicating the reference values for each variable that uses one.
It assumes that the same reference value is used whenever a variable appears in multiple terms in the \code{formula}
input of \code{hlnm}.}
}
\value{
A data.frame ready to use for plotting the results.
}
\description{
Replicates the loop in \code{hlnm} that construct the design matrices, without modifying \code{terms}
and creating the other objects (i.e., the precision matrices).
}
